feat: rename intern pkg to internal
