+++ cmd/convert/main.go
@@ -10,9 +10,9 @@ import (
 
 	bolt "go.etcd.io/bbolt"
 
-	"github.com/quixsi/core/intern/db"
-	"github.com/quixsi/core/intern/db/jsondb"
-	"github.com/quixsi/core/intern/db/kvdb"
+	"github.com/quixsi/core/internal/db"
+	"github.com/quixsi/core/internal/db/jsondb"
+	"github.com/quixsi/core/internal/db/kvdb"
 )
 
 func main() {
