+++ internal/db/jsondb/guest_store.go
@@ -14,7 +14,7 @@ import (
 	"github.com/google/uuid"
 	"go.opentelemetry.io/otel/trace"
 
-	"github.com/quixsi/core/intern/model"
+	"github.com/quixsi/core/internal/model"
 )
 
 // GuestStore is an implementation of the GuestStore interface
