+++ internal/db/jsondb/tracing.go
@@ -5,4 +5,4 @@ package jsondb
 
 import "go.opentelemetry.io/otel"
 
-var tracer = otel.GetTracerProvider().Tracer("github.com/quixsi/core/intern/db/jsondb")
+var tracer = otel.GetTracerProvider().Tracer("github.com/quixsi/core/internal/db/jsondb")
