+++ internal/db/kvdb/invitation_store.go
@@ -12,7 +12,7 @@ import (
 	bolt "go.etcd.io/bbolt"
 	"go.opentelemetry.io/otel/trace"
 
-	"github.com/quixsi/core/intern/model"
+	"github.com/quixsi/core/internal/model"
 )
 
 const bucketInvitation = "invitation_store"
