+++ internal/model/error.go
