+++ internal/model/event.go
