+++ internal/server/server.go
@@ -19,9 +19,9 @@ import (
 	"go.opentelemetry.io/otel/codes"
 	"go.opentelemetry.io/otel/trace"
 
-	"github.com/quixsi/core/intern/db"
-	"github.com/quixsi/core/intern/model"
-	"github.com/quixsi/core/intern/server/templates"
+	"github.com/quixsi/core/internal/db"
+	"github.com/quixsi/core/internal/model"
+	"github.com/quixsi/core/internal/server/templates"
 )
 
 //go:embed all:static
