+++ internal/server/static/lets-party.png
