+++ internal/server/templates/admin.event.html
