+++ internal/server/templates/admin.event.location.hotel.html
