+++ internal/server/templates/admin.event.location.html
