+++ internal/server/templates/admin.invitation-table-row.html
