+++ internal/server/templates/admin.nav.html
