+++ internal/server/templates/admin.translations.html
