+++ internal/server/templates/footer.html
