+++ internal/server/templates/guest-input.html
