+++ internal/server/templates/hotels.html
