+++ internal/server/templates/invitation.banner.html
