+++ internal/server/templates/invitation.content.html
