+++ internal/server/templates/invitation.nav.html
