+++ internal/server/templates/location.html
