+++ internal/server/templates/map.html
