+++ internal/server/templates/toast.error.html
