+++ internal/server/templates/toast.success.html
