+++ internal/server/templates/tracing.go
@@ -5,4 +5,4 @@ package templates
 
 import "go.opentelemetry.io/otel"
 
-var tracer = otel.GetTracerProvider().Tracer("github.com/quixsi/core/intern/server/templates")
+var tracer = otel.GetTracerProvider().Tracer("github.com/quixsi/core/internal/server/templates")
