+++ internal/server/tracing.go
@@ -1,8 +1,8 @@
 // Copyright (C) 2024 the lets-party maintainers
 // See root-dir/LICENSE for more information
 
-package kvdb
+package server
 
 import "go.opentelemetry.io/otel"
 
-var tracer = otel.GetTracerProvider().Tracer("github.com/quixsi/core/intern/db/kvdb")
+var tracer = otel.GetTracerProvider().Tracer("github.com/quixsi/core/internal/server")
